Use cyw43_wifi_get_rssi() for rssi updates (new in SDK 1.5.0).
- The return type is int32_t.

- In poll mode, cyw43_wifi_get_rssi() may not be called in an ISR,
  and must be called from the core on which networking was initialized.

The last condition means that rssi updates must be implemented
differently for poll mode. In background mode, we run the updates
from the repeating timer on core1, as before. In poll mode, the
repeating timer just sets a boolean to indicate the timeout, upon
which the update is run from the main loop on core0.

+++ src/handlers.h
@@ -40,10 +40,10 @@ struct netinfo {
 uint32_t get_temp(void);
 
 /*
- * Return the most recent rssi value for "our" access point, or UINT16_MAX
+ * Return the most recent rssi value for "our" access point, or INT32_MAX
  * if no rssi value has been read.
  */
-int16_t get_rssi(void);
+int32_t get_rssi(void);
 
 /*
  * Custom response handlers for the URL paths:
